Spliting functions from show to common for shared use in SF
