+++ .gitignore
@@ -16,6 +16,7 @@ install-sh
 missing
 stamp-*
 show
+sf
 *.exe
 *.stackdump
 *~
