+++ src/common.h
@@ -1,1 +1,4 @@
 void printVersion(char* programName);
+void printMenu(int line, int col, char *menustring);
+void topLineMessage(const char *message);
+int exittoshell();
