+++ src/menus.c
@@ -29,6 +29,7 @@
 #include <libgen.h>
 #include <signal.h>
 #include <regex.h>
+#include "common.h"
 #include "functions.h"
 #include "show.h"
 #include "views.h"
@@ -94,8 +95,6 @@ extern char *objectWild;
 
 //char testMenu[256];
 
-void topLineMessage(const char *message);
-
 int sanitizeTopFileRef(int topfileref)
 {
   if ( (selected - topfileref ) > displaysize ){
@@ -140,43 +139,6 @@ void refreshDirectory(char *sortmode, int origtopfileref, int origselected)
     topfileref = sanitizeTopFileRef(selected);
   }
   display_dir(currentpwd, ob, topfileref, selected);
-}
-
-void printMenu(int line, int col, char *menustring)
-{
-  int i, len, charcount;
-  charcount = 0;
-  move(line, col);
-  clrtoeol();
-  len = strlen(menustring);
-  setColors(COMMAND_PAIR);
-  for (i = 0; i < len; i++)
-     {
-      if ( menustring[i] == '!' ) {
-          i++;
-          setColors(HILITE_PAIR);
-          mvprintw(line, col + charcount, "%c", menustring[i]);
-          setColors(COMMAND_PAIR);
-          charcount++;
-      } else if ( menustring[i] == '<' ) {
-          i++;
-          setColors(HILITE_PAIR);
-          mvprintw(line, col + charcount, "%c", menustring[i]);
-          charcount++;
-      } else if ( menustring[i] == '>' ) {
-          i++;
-          setColors(COMMAND_PAIR);
-          mvprintw(line, col + charcount, "%c", menustring[i]);
-          charcount++;
-      } else if ( menustring[i] == '\\' ) {
-          i++;
-          mvprintw(line, col + charcount, "%c", menustring[i]);
-          charcount++;
-      } else {
-          mvprintw(line, col + charcount, "%c", menustring[i]);
-          charcount++;
-        }
-    }
 }
 
 void directory_view_menu_inputs(); // Needed to allow menu inputs to switch between each other
@@ -1271,6 +1233,7 @@ void global_menu_inputs()
           break;
         case 'q':
           if (historyref == 0){
+            free(hs);
             exittoshell();
             refresh();
           } else {
@@ -1289,22 +1252,3 @@ void global_menu_inputs()
         }
     }
 }
-
-void topLineMessage(const char *message){
-  move(0,0);
-  clrtoeol();
-  setColors(ERROR_PAIR);
-  mvprintw(0,0, "%s", message);
-  setColors(COMMAND_PAIR);
-  while(1)
-    {
-      *pc = getch();
-      switch(*pc)
-        {
-        default: // Where's the "any" key?
-          directory_view_menu_inputs();
-          break;
-        }
-    }
-}
-
