+++ src/menus.h
@@ -16,8 +16,6 @@ void delete_multi_file_confirm_input(results* ob);
 void copy_multi_file_input(results* ob, char *input);
 void copy_multi_file(results* ob, char *dest);
 void rename_multi_file_input(results* ob, char *input);
-void topLineMessage(const char *message);
-void printMenu(int line, int col, char *menustring);
 int huntCaseSelectInput();
 void huntInput(int selected, int charcase);
 void refreshDirectory(char *sortmode, int topfileref, int selected);
